SAN-518 Refactor ViewPenaltiesController
